Added LiveChat.js
Move code from room.js to LiveChat.js für mehr Ordnung
