+++ app/resources/js/room.js
@@ -1,21 +1,20 @@
+/* eslint-env node */
+
+import {LiveChat} from "./LiveChat.js";
+
 let nicknameTextField,
-messageTextField;
+liveChatInstance;
 
 function init(){
     setClickListener();
+    liveChatInstance = new LiveChat();
 }
 
 function setClickListener(){
-    messageTextField = document.querySelector(".typeField");
     nicknameTextField = document.querySelector(".nickname");
     nicknameTextField.addEventListener("keypress", function (e){
         if(e.key === "Enter" && nicknameTextField.value !== ""){
             enterNickname();
-        }
-    });
-    messageTextField.addEventListener("keypress", function (e){
-        if(e.key === "Enter" && messageTextField.value !== ""){
-            addMessage();
         }
     });
 }
@@ -24,13 +23,6 @@ function enterNickname(){
     nicknameTextField.classList.add("hidden");
     document.querySelector(".chat").classList.remove("disabled");
     document.querySelector(".typeField").disabled = false;
-}
-
-function addMessage(){
-    var chat = document.querySelector(".chat"),
-    message = document.createElement("div");
-    message.innerHTML = messageTextField.value;
-    chat.insertBefore(message, document.querySelector(".typeField"));
 }
 
 init();
